+++ src/shape/groups/common.js
@@ -10,10 +10,13 @@ const Global = require('../../global');
 Shape.registerGroup('common', {
   draw(item) {
     const model = item.getModel();
+    let keyShape;
     if (model.collapsed) {
-      return this.drawCollapsed(item);
+      keyShape = this.drawCollapsed(item);
+    } else {
+      keyShape = this.drawExpanded(item);
     }
-    return this.drawExpanded(item);
+    return keyShape;
   },
   defaultWidth: 184,
   defaultHeight: 40,
@@ -78,6 +81,7 @@ Shape.registerGroup('common', {
     const box = {
       ...item.lastChildrenBox
     };
+    const model = item.getModel();
     if (item.getChildren().length > 0) {
       const childrenBBox = item.getChildrenBBox();
       box.x = childrenBBox.minX - Global.groupBackgroundPadding[3];
@@ -87,6 +91,12 @@ Shape.registerGroup('common', {
     } else {
       box.width = this.defaultWidth;
       box.height = this.defaultHeight;
+    }
+    if (Util.isNil(box.x) && !Util.isNil(model.x)) {
+      box.x = model.x;
+    }
+    if (Util.isNil(box.y) && !Util.isNil(model.y)) {
+      box.y = model.y;
     }
     return box;
   },